Move functions from utils to libs
