+++ src/libs/slateOpsToAutomerge.js
